added logic to avoid making redundant requests

+++ retrieve_full_texts.sh
@@ -17,18 +17,20 @@ while read pmc_id; do
 	request_url="$PREFIX$pmc_id&api_key=$NCBI_API_KEY$SUFFIX"
 	
 	output_path=$(echo -n $pmc_id | sed -E 's#(\S+)#pmc_xmls/\1.xml#')
+	
+	export elapsed="0"
 
-	# TODO: need to check if file exists if doing potentially 
-	# redundant runs
-
-	# track request time
-	req_start=$(date +%s.%N)
-	wget -o pull.errs -nv -O $output_path $request_url
-	elapsed=$(date +%s.%N --date="$req_start seconds ago")
+	# Avoid making redundant requests
+	if !([ -e $output_path ]); then
+		# track request time
+		req_start=$(date +%s.%N)
+		wget -o pull.errs -nv -O $output_path $request_url
+		elapsed=$(date +%s.%N --date="$req_start seconds ago")
+	fi
 
 	# avoid making more than 10 requests per second
 	if (( $(echo "$elapsed<0.1" | bc -l) )); then
 		wait_time=$(echo "0.1-$elapsed" | bc)
 		sleep $wait_time
 	fi
-done < pmc_ids
+done < selected_articles
